refactoring initialization
