+++ si-id-epu50/MANIFEST.in
@@ -1,0 +1,1 @@
+include si_id_epu50/access_rules.as
