+++ si-id-epu50/requirements.txt
@@ -1,5 +1,5 @@
-pyyaml=6.0
-toml=0.10.2
-pydantic=1.10.2
-pcaspy=0.7.3
-pyepics=3.5.0
+pyyaml>=6.0
+toml>=0.10.2
+pydantic>=1.10.2
+pcaspy>=0.7.3
+pyepics>=3.5.0
