+++ si-id-epu50/si_id_epu50/autosave_epu.req
@@ -1,0 +1,5 @@
+# speed set point
+GapSpeed-SP
+PhaseSpeed-SP
+MaxGapSpeed-SP
+MaxPhaseSpeed-SP
