+++ si-id-epu50/source/__init__.py
@@ -1,10 +1,1 @@
-"""EPU IOC package."""
-
-from . import constants
-from . import epu_db
-from . import iocDriver
-from . import utils
-from . import ecodrive_ioc
-from . import ecodrive
-from . import epu
-from . import save_restore
+../si_id_epu50/__init__.py
