+++ si-id-epu50/source/autosave_epu.req
@@ -1,5 +1,1 @@
-# speed set point
-GapSpeed-SP
-PhaseSpeed-SP
-MaxGapSpeed-SP
-MaxPhaseSpeed-SP
+../si_id_epu50/autosave_epu.req
