+++ si-id-epu50/source/epu.as
@@ -1,15 +1,1 @@
-# IOC ACCESS SECURITY FILE
-
-# full write and read access
-# 
-ASG(default) {
-    RULE(1, READ)
-    RULE(1, WRITE)
-}
-
-# read-only PVs
-# 
-ASG(readonly) {
-    RULE(1, READ)
-    RULE(0, WRITE)
-}
+../si_id_epu50/access_rules.as
